emitNotif clean up
