+++ packages/commonwealth/server/util/emitNotifications/emitChainEventNotifications.ts
@@ -1,5 +1,5 @@
 import { StatsDController } from 'common-common/src/statsd';
-import Sequelize, { QueryTypes } from 'sequelize';
+import Sequelize from 'sequelize';
 import type { IChainEventNotificationData } from 'types';
 import { SERVER_URL } from '../../config';
 import type { DB } from '../../models';
@@ -8,64 +8,21 @@ import {
   createImmediateNotificationEmailObject,
   sendImmediateNotificationEmail,
 } from '../../scripts/emails';
-import type { WebhookContent } from '../../webhookNotifier';
 import { factory, formatFilename } from 'common-common/src/logging';
 import { NotificationCategories } from 'common-common/src/types';
+import { filterAddresses } from './util';
 
 const log = factory.getLogger(formatFilename(__filename));
 
 const { Op } = Sequelize;
 
-export async function filterAddresses(
-  models,
-  findOptions: any,
-  includedAddresses: string[],
-  excludedAddresses: string[]
-) {
-  const query = `
-        SELECT DISTINCT user_id
-        FROM "Addresses"
-        WHERE address IN (?);
-      `;
-  // currently excludes override includes, but we may want to provide the option for both
-  if (excludedAddresses && excludedAddresses.length > 0) {
-    const ids = <number[]>(
-      (<unknown>(
-        await models.sequelize.query(query, {
-          type: QueryTypes.SELECT,
-          raw: true,
-          replacements: excludedAddresses,
-        })
-      ))
-    );
-    if (ids && ids.length > 0) {
-      findOptions[Op.and].push({ subscriber_id: { [Op.notIn]: ids } });
-    }
-  } else if (includedAddresses && includedAddresses.length > 0) {
-    const ids = <number[]>(
-      (<unknown>(
-        await models.sequelize.query(query, {
-          type: QueryTypes.SELECT,
-          raw: true,
-          replacements: includedAddresses,
-        })
-      ))
-    );
-
-    if (ids && ids.length > 0) {
-      findOptions[Op.and].push({ subscriber_id: { [Op.in]: ids } });
-    }
-  }
-  return findOptions;
-}
-
 export default async function emitChainEventNotification(
   models: DB,
-  chain_id: string,
   notification_data: IChainEventNotificationData,
   excludeAddresses?: string[],
   includeAddresses?: string[]
 ): Promise<NotificationInstance> {
+  const chain_id = notification_data.chain;
   // get subscribers to send notifications to
   StatsDController.get().increment('cw.notifications.created', {
     category_id: NotificationCategories.ChainEvent,
@@ -123,11 +80,13 @@ export default async function emitChainEventNotification(
   }
 
   // send emails
-  for (const subscription of subscriptions) {
-    if (msg && subscription?.immediate_email) {
-      // TODO: verify this works
-      msg.dynamic_template_data.notification.path = `${SERVER_URL}/${notification_data.chain}/notifications?id=${notification.id}`;
-      sendImmediateNotificationEmail(subscription.User, msg);
+  if (msg) {
+    for (const subscription of subscriptions) {
+      if (subscription?.immediate_email) {
+        // TODO: verify this works
+        msg.dynamic_template_data.notification.path = `${SERVER_URL}/${notification_data.chain}/notifications?id=${notification.id}`;
+        sendImmediateNotificationEmail(subscription.User, msg);
+      }
     }
   }
